Don't rely on external sources to resolve voterid.
The external sources like electoralsearch.in doesn't have a consistent
way to find polling booth from voterid. They provide pb name, but that
is hard to resolve. For some states, the "id" field seems to provide
that info, but that is not always the case.

Using a local voter db that maintains mapping from voterid to its
polling booth.