Fixed the order of parents.
The parents were earlier order by place.id. That used to work because the
places that are top the hierarchy were added before the bottom ones. After
adding some places in the middle in the hierarchy, it started to break.

Changed the order to place.place_type_id. This is good enough for now as
place_types are not going to change that often.